refactor: change to ts
