+++ packages/loader/package.json
@@ -3,5 +3,6 @@
   "version": "1.3.6",
   "description": "",
   "license": "Apache-2.0",
-  "main": "./define-loader.mjs"
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts"
 }
